feat: player-specific extended info documentation

+++ protocol/osrs-221-desktop/src/main/kotlin/net/rsprot/protocol/game/outgoing/codec/playerinfo/extendedinfo/PlayerFaceAngleEncoder.kt
@@ -17,7 +17,7 @@ public class PlayerFaceAngleEncoder : PrecomputedExtendedInfoEncoder<FaceAngle> 
             alloc
                 .buffer(2, 2)
                 .toJagByteBuf()
-        buffer.p2(extendedInfo.angle)
+        buffer.p2(extendedInfo.angle.toInt())
         return buffer
     }
 }
